requirements.txt
cupy-cuda115==9.6.0
decorator==4.4.2
Flask==2.0.3
Jinja2==3.0.3
matplotlib==3.1.1
numba==0.47.0
numpy==1.16.4
opencv-python==4.5.5.62
pandas==0.25.3
Pillow==7.2.0
requests==2.22.0
requests-oauthlib==1.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
tensorflow==2.0.0
tensorflow-estimator==1.14.0
